Added support for the staticman comment engine
In the postfooter I added conditional check to include staticman
partials if the staticman configuration is present. If the staticman
configuration is not present then the disqus template is used.

Added a create and view partial for staticman comments.

updated the layout.css to include flex-box styles to make form in the
staticman create-comments partial render nicely.